refactor: 안쓰는 코드 제거 및 디렉토리 이동
