+++ src/main/java/com/pawland/global/config/security/handler/LoginSuccessHandler.java
@@ -2,7 +2,7 @@ package com.pawland.global.config.security.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.pawland.global.config.security.JwtUtils;
-import com.pawland.global.config.security.UserPrincipal;
+import com.pawland.global.config.security.domain.UserPrincipal;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -27,14 +27,15 @@ public class LoginSuccessHandler implements AuthenticationSuccessHandler {
 
     private final ObjectMapper objectMapper;
     private final JwtUtils jwtUtils;
+    private static final String JWT_NAME = "jwt";
 
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
         UserPrincipal principal = (UserPrincipal) authentication.getPrincipal();
         String jwt = jwtUtils.generateAccessToken(principal.getUsername(), new Date());
 
-        ResponseCookie cookie = ResponseCookie.from("jwt", jwt)
-            .domain("localhost")   // todo 서버 환경에 따라 설정파일로 분리해서 관리
+        ResponseCookie cookie = ResponseCookie.from(JWT_NAME, jwt)
+            .domain("localhost")   // TODO: 서버 환경에 따라 설정파일로 분리해서 관리
             .path("/")
             .secure(false)
             .maxAge(Duration.ofDays(30))
@@ -45,6 +46,6 @@ public class LoginSuccessHandler implements AuthenticationSuccessHandler {
         response.setCharacterEncoding(UTF_8.name());
         response.setStatus(SC_OK);
         response.setHeader(HttpHeaders.SET_COOKIE, cookie.toString());
-        objectMapper.writeValue(response.getWriter(), "로그인에 성공했습니다.");
+        objectMapper.writeValue(response.getWriter(), "로그인에 성공했습니다."); // TODO: 메시지 Enum으로 관리
     }
 }
